testing: move swarm related handlers to another file
